+++ packages/gestalt/rollup.config.js
@@ -1,18 +1,21 @@
-import path from 'pathe';
-import fg from 'fast-glob';
+import path from 'pathe'
+import fg from 'fast-glob'
 
-import {external, plugins, distDir} from '../../configurations/rollup.config';
+import { external, plugins, distDir } from '../../configurations/rollup.config'
 
-const gestaltExternal = [...external, '@oclif/core', '@gestaltjs/core/cli', "@gestaltjs/core/framework"]
-const gestaltPlugins = [
-  ...plugins(__dirname),
+const gestaltExternal = [
+  ...external,
+  '@oclif/core',
+  '@gestaltjs/core/cli',
+  '@gestaltjs/core/framework',
 ]
-const gestaltFeatures = ['build', 'db', 'lint', 'serve', 'test', 'type-check'];
+const gestaltPlugins = [...plugins(__dirname)]
+const gestaltFeatures = ['build', 'db', 'lint', 'serve', 'test', 'check']
 const gestaltCommands = gestaltFeatures.flatMap((feature) => {
   return fg.sync([
     path.join(__dirname, `../${feature}/src/cli/commands/**/*.ts`),
     `!${path.join(__dirname, `../${feature}/src/cli/commands/**/*.test.ts`)}`,
-  ]);
+  ])
 })
 
 const configuration = () => [
@@ -28,16 +31,16 @@ const configuration = () => [
             return `commands/${chunkInfo.facadeModuleId
               .split('src/cli/commands')
               .slice(-1)[0]
-              .replace('ts', 'js')}`;
+              .replace('ts', 'js')}`
           } else {
-            return '[name].js';
+            return '[name].js'
           }
         },
       },
     ],
     plugins: gestaltPlugins,
     external: gestaltExternal,
-  }
-];
+  },
+]
 
-export default configuration;
+export default configuration